rename: model package 생성 후 calculator class 이동
