+++ src/main/java/calculator/model/Calculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.model;
 
 import java.util.List;
